--- a/semanticdb-javac/src/main/java/com/sourcegraph/semanticdb_javac/LocalSymbolsCache.java
+++ b/semanticdb-javac/src/main/java/com/sourcegraph/semanticdb_javac/LocalSymbolsCache.java
@@ -1,20 +1,20 @@
 package com.sourcegraph.semanticdb_javac;
 
-import com.sun.tools.javac.code.Symbol;
+import javax.lang.model.element.Element;
 
 import java.util.IdentityHashMap;
 
 /** Cache of SemanticDB symbols that are local to a single file. */
 public final class LocalSymbolsCache {
 
-  private final IdentityHashMap<Symbol, String> symbols = new IdentityHashMap<>();
+  private final IdentityHashMap<Element, String> symbols = new IdentityHashMap<>();
   private int localsCounter = -1;
 
-  public String get(Symbol sym) {
+  public String get(Element sym) {
     return symbols.get(sym);
   }
 
-  public String put(Symbol sym) {
+  public String put(Element sym) {
     localsCounter++;
     String result = SemanticdbSymbols.local(localsCounter);
     symbols.put(sym, result);