--- a/semanticdb-javac/src/main/java/com/sourcegraph/semanticdb_javac/RangeFinder.java
+++ b/semanticdb-javac/src/main/java/com/sourcegraph/semanticdb_javac/RangeFinder.java
@@ -1,60 +1,93 @@
 package com.sourcegraph.semanticdb_javac;
 
-import com.sun.source.tree.CompilationUnitTree;
-import com.sun.source.tree.LineMap;
-import com.sun.source.util.TreePath;
-import com.sun.source.util.Trees;
+import javax.tools.Diagnostic;
 
 import javax.lang.model.element.Element;
-import javax.lang.model.element.Name;
-import java.io.IOException;
-import java.util.HashMap;
 import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import static com.sourcegraph.semanticdb_javac.Debugging.pprint;
 
 public class RangeFinder {
-  public static Optional<Semanticdb.Range> findRange(
-      TreePath path,
-      Trees trees,
-      CompilationUnitTree root,
+  public static class StartEndRange {
+    public int start;
+    public int end;
+
+    StartEndRange(int start, int end) {
+      this.start = start;
+      this.end = end;
+    }
+  }
+
+  public static Optional<StartEndRange> findRange(
       Element element,
-      int startPos,
+      String name,
+      int originalStartPos,
+      int originalEndPos,
       String source,
       boolean fromEnd) {
-    LineMap lineMap = root.getLineMap();
-    Name name = element.getSimpleName();
-    if (name.contentEquals("<init>")) name = element.getEnclosingElement().getSimpleName();
-
-    int endPos = (int) trees.getSourcePositions().getEndPosition(root, path.getLeaf());
-    // false for anonymous classes
-    if (name.length() != 0) {
-      startPos = findNameIn(name, fromEnd ? endPos : startPos, source, fromEnd);
-      endPos = startPos + name.length();
-    }
+    int startPos = findNameIn(name, originalStartPos, originalEndPos, element, source, fromEnd);
+    int endPos = startPos + name.length();
 
     if (endPos == -1 || startPos == -1) {
       return Optional.empty();
     }
 
-    Semanticdb.Range range =
-        Semanticdb.Range.newBuilder()
-            .setStartLine((int) lineMap.getLineNumber(startPos) - 1)
-            .setStartCharacter((int) lineMap.getColumnNumber(startPos) - 1)
-            .setEndLine((int) lineMap.getLineNumber(endPos) - 1)
-            .setEndCharacter((int) lineMap.getColumnNumber(endPos) - 1)
-            .build();
-    return Optional.of(range);
+    return Optional.of(new StartEndRange(startPos, endPos));
   }
 
-  private static int findNameIn(CharSequence name, int start, String source, boolean fromEnd) {
-    if (source.equals("")) return -1;
+  private static int findNameFromEnd(
+      String name,
+      int originalStartPos,
+      int originalEndPos,
+      int end,
+      Element element,
+      String source) {
+    if (end < 0) return -1;
+    int offset = source.lastIndexOf(name, end);
+    if (offset == -1 && originalStartPos != Diagnostic.NOPOS && originalEndPos != Diagnostic.NOPOS)
+      return originalStartPos;
+    if (offset == -1) {
+      return -1;
+    }
+    int endOfWord = offset + name.length();
+    // found name in wrong word? e.g. finding `"A"` in `A("A")`
+    if (offset > 0 && Character.isJavaIdentifierPart(source.charAt(offset - 1)))
+      return findNameFromEnd(name, originalStartPos, originalEndPos, offset - 1, element, source);
+    if (endOfWord < source.length() && Character.isJavaIdentifierPart(source.charAt(endOfWord)))
+      return findNameFromEnd(name, originalStartPos, originalEndPos, offset - 1, element, source);
+
+    return offset;
+  }
+
+  private static int findNameFromStart(
+      String name,
+      int start,
+      int originalStartPos,
+      int originalEndPos,
+      Element element,
+      String source) {
+    if (start >= source.length()) return -1;
+    int offset = source.indexOf(name, start);
+    if (offset == -1 && originalStartPos != Diagnostic.NOPOS && originalEndPos != Diagnostic.NOPOS)
+      return originalStartPos;
+    if (offset == -1) {
+      return -1;
+    }
+    int end = offset + name.length();
+    // found name in wrong word? e.g. finding `"A"` in `A("A")`
+    if (offset > 0 && Character.isJavaIdentifierPart(source.charAt(offset - 1)))
+      return findNameFromStart(name, end + 1, originalStartPos, originalEndPos, element, source);
+    if (end < source.length() && Character.isJavaIdentifierPart(source.charAt(end)))
+      return findNameFromStart(name, end + 1, originalStartPos, originalEndPos, element, source);
+
+    return offset;
+  }
+
+  private static int findNameIn(
+      String name, int start, int end, Element element, String source, boolean fromEnd) {
+    if (source.length() == 0) return -1;
 
     int offset;
-    if (fromEnd) offset = source.lastIndexOf(name.toString(), start);
-    else offset = source.indexOf(name.toString(), start);
+    if (fromEnd) offset = findNameFromEnd(name, start, start, end, element, source);
+    else offset = findNameFromStart(name, start, end, end, element, source);
     if (offset > -1) {
       return offset;
     }