This is an attempt to use the JDK public API instead of the private one.

The advantage is that for JDK 17+ there should be (mostly) no need for `--add-exports` which should make it easier for tools that use this plugin e.g. Metals.

I can't test this completely locally as I'm using Windows and some tests won't run.
I'm also not entirely sure what the outputs _should_ be.  In some cases they look better e.g. `enums` and for some parameterised types.
For the library tests there are a lot of differences but the correct packages seem to now be displayed in test results instead of `_root_`.

I have no idea how this affects performance.  `Types`, `Trees` and `Elements` now sometimes have to be queried to lookup data instead of being able to query the internal classes directly so maybe slower 🤷 

I've changed the `-targetroot:javac-classes-directory` option to use reflection so `--add-exports jdk.compiler/com.sun.tools.javac.util=ALL-UNNAMED` only has to be supplied if that option is specified.

### Test plan

Rerun all the tests I guess