@Arthurm1 2023-07-25
@olafurpg I've reverted the `overrides` search to use the internal API.  I've also switched to regex split for package names.

I've also managed to run the tests under WSL.

I guess I could fix the enums `<init>` reference by creating an `<init>` definition on enum itself - as is done with classes that don't have constructors.

This should be good enough to merge though.

I've highlighted internal API usage with warning comments.  Is it possible to test the performance of the `overrides` code?  I've left the public API version in the file but commented it out, so switching between public and private implementations should be easy for whoever is able to benchmark it.  There is another option for implementing using the public API if this one sucks.