@olafurpg 2023-07-25
Thank you @Arthurm1 ! The repo has JMH benchmarks that we can use to confirm performance claims. I will review the changes tomorrow and share instructions on how to run the benchmarks.