@olafurpg 2023-07-26
To fix the override suite, you need to manually update the string literal in the assertion. There's no automatic way to update that snapshot