@Arthurm1 2023-07-26
@olafurpg

Using the public API is about 3% slower if I'm reading these results right? I'm no expert on benchmarking

Private API
```
[info] Result "benchmarks.CompileBench.compileSemanticdb":
[info]   N = 10
[info]   mean =  16387.775 ±(99.9%) 281.073 ms/op
[info]   Histogram, ms/op:
[info]     [16100.000, 16150.000) = 0 
[info]     [16150.000, 16200.000) = 2 
[info]     [16200.000, 16250.000) = 1 
[info]     [16250.000, 16300.000) = 0 
[info]     [16300.000, 16350.000) = 1 
[info]     [16350.000, 16400.000) = 2 
[info]     [16400.000, 16450.000) = 0 
[info]     [16450.000, 16500.000) = 2 
[info]     [16500.000, 16550.000) = 0 
[info]     [16550.000, 16600.000) = 0 
[info]     [16600.000, 16650.000) = 1 
[info]     [16650.000, 16700.000) = 0 
[info]     [16700.000, 16750.000) = 1
[info]   Percentiles, ms/op:
[info]       p(0.0000) =  16151.304 ms/op
[info]      p(50.0000) =  16379.456 ms/op
[info]      p(90.0000) =  16724.546 ms/op
[info]      p(95.0000) =  16737.928 ms/op
[info]      p(99.0000) =  16737.928 ms/op
[info]      p(99.9000) =  16737.928 ms/op
[info]      p(99.9900) =  16737.928 ms/op
[info]      p(99.9990) =  16737.928 ms/op
[info]      p(99.9999) =  16737.928 ms/op
[info]     p(100.0000) =  16737.928 ms/op
[info] # Run complete. Total time: 00:09:34
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                       (lib)  Mode  Cnt      Score     Error  Units
[info] CompileBench.compile            guava    ss   10  10235.191 ± 239.764  ms/op
[info] CompileBench.compileSemanticdb  guava    ss   10  16387.775 ± 281.073  ms/op
[success] Total time: 585 s (09:45), completed Jul 26, 2023, 10:52:52 AM
```

Public API
```
[info] Result "benchmarks.CompileBench.compileSemanticdb":
[info]   N = 10
[info]   mean =  16835.790 ±(99.9%) 166.706 ms/op
[info]   Histogram, ms/op:
[info]     [16600.000, 16650.000) = 0 
[info]     [16650.000, 16700.000) = 2 
[info]     [16700.000, 16750.000) = 0 
[info]     [16750.000, 16800.000) = 1 
[info]     [16800.000, 16850.000) = 1 
[info]     [16850.000, 16900.000) = 4 
[info]     [16900.000, 16950.000) = 1 
[info]     [16950.000, 17000.000) = 0 
[info]     [17000.000, 17050.000) = 1 
[info]   Percentiles, ms/op:
[info]       p(0.0000) =  16653.959 ms/op
[info]      p(50.0000) =  16853.743 ms/op
[info]      p(90.0000) =  17013.641 ms/op
[info]      p(95.0000) =  17026.258 ms/op
[info]      p(99.0000) =  17026.258 ms/op
[info]      p(99.9000) =  17026.258 ms/op
[info]      p(99.9900) =  17026.258 ms/op
[info]      p(99.9990) =  17026.258 ms/op
[info]      p(99.9999) =  17026.258 ms/op
[info]     p(100.0000) =  17026.258 ms/op
[info] # Run complete. Total time: 00:09:38
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                       (lib)  Mode  Cnt      Score     Error  Units
[info] CompileBench.compile            guava    ss   10   9931.021 ± 134.260  ms/op
[info] CompileBench.compileSemanticdb  guava    ss   10  16835.790 ± 166.706  ms/op
[success] Total time: 585 s (09:45), completed Jul 26, 2023, 11:51:50 AM
```

You should be able to merge this PR now as tests pass