@Arthurm1 2023-08-04
Reworked to cache all Trees in a Map until full scan is complete.  Then tree references can be looked up directly in the Map (which caused a lot of the performance drop off).

Now performance is on par with current release...

```
Private API (current release)
[info] Benchmark                       (lib)  Mode  Cnt      Score     Error  Units
[info] CompileBench.compile            guava    ss   10   9915.177 ± 189.343  ms/op
[info] CompileBench.compileSemanticdb  guava    ss   10  12442.871 ± 295.333  ms/opn

Public API (this PR)
[info] Benchmark                       (lib)  Mode  Cnt      Score     Error  Units
[info] CompileBench.compile            guava    ss   10  10338.916 ± 246.845  ms/op
[info] CompileBench.compileSemanticdb  guava    ss   10  12200.354 ± 716.330  ms/op
```