@keynmol 2023-08-08
Hey @Arthurm1 sorry for radio silence on this PR - I will take a look and re-run benchmarks just to reproduce what you are seeing.

Exciting times!