@Arthurm1 2023-08-08
@keynmol Sorry - accidentally included some notes so pushed twice

All done I think - you should be able to test performance.