@keynmol 2023-08-16
@Arthurm1 This morning I ran the benchmarks on my linux machine, glanced at them, put them in a file and... left the house without pasting the results 🤦 

From my quick glance I could confirm the performance was on part over 10 warmup iterations and 10 iterations. I'll review the PR with that in mind, and will paste the results later today to make sure there's no regression.