@keynmol 2023-08-16
I finally got back to that machine - benchmarks look good!

```
main:

[info] CompileBench.compile              bytebuddy    ss   10  2293.563 ± 147.362  ms/op
[info] CompileBench.compile                  guava    ss   10  3065.379 ± 107.811  ms/op
[info] CompileBench.compileSemanticdb    bytebuddy    ss   10  2871.277 ± 109.309  ms/op
[info] CompileBench.compileSemanticdb        guava    ss   10  3856.719 ±  61.454  ms/op
[info] ScipSemanticdbBench.json                N/A    ss   10  3154.109 ±  41.637  ms/op
[info] ScipSemanticdbBench.jsonParallel        N/A    ss   10  2208.542 ±  53.408  ms/op

This PR

[info] CompileBench.compile              bytebuddy    ss   10  2294.170 ± 97.459  ms/op
[info] CompileBench.compile                  guava    ss   10  2921.679 ± 37.208  ms/op
[info] CompileBench.compileSemanticdb    bytebuddy    ss   10  3016.769 ± 35.410  ms/op
[info] CompileBench.compileSemanticdb        guava    ss   10  3789.773 ± 76.272  ms/op
[info] ScipSemanticdbBench.json                N/A    ss   10  3144.038 ± 33.890  ms/op
[info] ScipSemanticdbBench.jsonParallel        N/A    ss   10  2226.563 ± 60.759  ms/op
```

I'll take one more look tomorrow morning and we'll merge and release this